Restructure for MVC
- restructure to better match MVC pattern
