+++ app_server/controllers/index.js
@@ -1,0 +1,7 @@
+const index = function(req, res, next) {
+    res.render('index', { title: 'Express' });
+};
+
+module.exports = {
+    index
+};
