+++ app_server/routes/index.js
@@ -1,9 +1,9 @@
 var express = require('express');
 var router = express.Router();
 
+const ctrlIndex = require('../controllers/index');
+
 /* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express' });
-});
+router.get('/', ctrlIndex.index);
 
 module.exports = router;
