+++ app_server/views/index/index.pug
@@ -1,4 +1,4 @@
-extends layout
+extends ../layout
 
 block content
   h1= title
